Moved roslaunch to common. But circular dependencies.
