+++ setup.cfg
@@ -54,12 +54,13 @@ per-file-ignores =
   src/roswire/description/__init__.py:F401
   src/roswire/parameters/__init__.py:F401
   src/roswire/proxy/__init__.py:F401
-  src/roswire/proxy/roslaunch/__init__.py:F401
-  src/roswire/proxy/roslaunch/config/__init__.py:F401
-  src/roswire/proxy/roslaunch/config/env.py:N817
-  src/roswire/proxy/roslaunch/config/node.py:N817
-  src/roswire/proxy/roslaunch/config/launch.py:N817
-  src/roswire/proxy/roslaunch/config/parameter.py:N817
+  src/roswire/common/__init__.py:F401
+  src/roswire/common/roslaunch/__init__.py:F401
+  src/roswire/common/roslaunch/config/__init__.py:F401
+  src/roswire/common/roslaunch/config/env.py:N817
+  src/roswire/common/roslaunch/config/node.py:N817
+  src/roswire/common/roslaunch/config/launch.py:N817
+  src/roswire/common/roslaunch/config/parameter.py:N817
   src/roswire/ros1/__init__.py:F401
   src/roswire/ros1/bag/__init__.py:F401
   src/roswire/ros1/roslaunch/__init__.py:F401
