+++ src/roswire/common/roslaunch/config/launch.py
