+++ src/roswire/common/roslaunch/config/node.py
