+++ src/roswire/proxy/__init__.py
@@ -1,3 +1,2 @@
 # -*- coding: utf-8 -*-
 from .ros1 import ROS1
-from .roslaunch import ROSLaunchManager
