+++ src/roswire/ros1/roslaunch/reader.py
@@ -12,13 +12,13 @@ import attr
 import dockerblade
 from loguru import logger
 
+from ...common.roslaunch.config import ExecutableType, LaunchConfig, NodeConfig
+from ...common.roslaunch.context import LaunchContext
+from ...common.roslaunch.reader import LaunchFileReader
+from ...common.roslaunch.rosparam import load_from_yaml_string as load_rosparam_from_string
+from ...common.roslaunch.substitution import ArgumentResolver
 from ...exceptions import FailedToParseLaunchFile
 from ...name import global_name, name_is_global, name_is_private, namespace, namespace_join
-from ...proxy.roslaunch.config import ExecutableType, LaunchConfig, NodeConfig
-from ...proxy.roslaunch.context import LaunchContext
-from ...proxy.roslaunch.reader import LaunchFileReader
-from ...proxy.roslaunch.rosparam import load_from_yaml_string as load_rosparam_from_string
-from ...proxy.roslaunch.substitution import ArgumentResolver
 
 if typing.TYPE_CHECKING:
     from ... import AppInstance
@@ -430,7 +430,8 @@ class ROS1LaunchFileReader(LaunchFileReader):
                             attrib: str,
                             ctx: LaunchContext,
                             default: None
-                            ) -> Optional[bool]: ...
+                            ) -> Optional[bool]:
+        ...
 
     @overload
     def _read_optional_bool(self,
@@ -438,7 +439,8 @@ class ROS1LaunchFileReader(LaunchFileReader):
                             attrib: str,
                             ctx: LaunchContext,
                             default: bool
-                            ) -> bool: ...
+                            ) -> bool:
+        ...
 
     def _read_optional_bool(self,
                             elem: ET.Element,
@@ -457,7 +459,8 @@ class ROS1LaunchFileReader(LaunchFileReader):
                              attrib: str,
                              ctx: LaunchContext,
                              default: None
-                             ) -> Optional[float]: ...
+                             ) -> Optional[float]:
+        ...
 
     @overload
     def _read_optional_float(self,
@@ -465,7 +468,8 @@ class ROS1LaunchFileReader(LaunchFileReader):
                              attrib: str,
                              ctx: LaunchContext,
                              default: float
-                             ) -> float: ...
+                             ) -> float:
+        ...
 
     def _read_optional_float(self,
                              elem: ET.Element,
